requirements.txt
certifi==2023.11.17
charset-normalizer==3.3.2
colorama==0.4.6
idna==3.6
loguru==0.7.2
numpy==1.26.2
opencv-python==4.8.1.78
Pillow==10.1.0
pypng==0.20220715.0
pyTelegramBotAPI==4.14.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
pyzbar==0.1.9
qrcode==7.4.2
requests==2.31.0
six==1.16.0
telebot==0.0.5
typing_extensions==4.9.0
tzdata==2023.3
urllib3==2.1.0
win32-setctime==1.1.0
